requirement.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Limiter==1.4
google-auth==1.27.1
google-auth-oauthlib==0.4.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
limits==1.5.1
MarkupSafe==1.1.1
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.15.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
